pyproject.toml
[project]
name = "web-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Aasher Kamal", email = "aasherkamal786@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "google-adk>=0.1.0",
]

[project.scripts]
web-agent = "web_agent:agent"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
